testheroku/requirements/production.txt
-r common.txt
gunicorn==19.9.0
dj-database-url==0.5.0
boto3==1.9.210
django-storages==1.7.1
